Get Celery working with Flask app context, other
