+++ services/document-manager/backend/Dockerfile
@@ -13,7 +13,7 @@ RUN pip install --no-cache-dir -r requirements.txt
 
 COPY . .
 
-# RUN celery worker -A app.utils.flask_celery --concurrency=8 -l info -f celery.log &
+# RUN celery worker -A app.utils.tasks --concurrency=8 -l info -f celery.log &
 
 # Run the server
 EXPOSE 5001
