+++ services/document-manager/backend/app/__init__.py
@@ -20,6 +20,8 @@ from app.extensions import api, cache, db, jwt, apm, migrate
 
 from .config import Config
 
+from celery import Celery
+
 
 def create_app(config_object=None):
     """Create and configure an instance of the Flask application."""
@@ -62,6 +64,26 @@ def register_extensions(app):
     jwt.init_app(app)
     migrate.init_app(app, db)
     CORS(app)
-    # celery.init_app(app)
 
     return None
+
+
+def make_celery(app=None):
+    app = app or create_app()
+    celery = Celery(
+        app.import_name,
+        backend=app.config['CELERY_RESULT_BACKEND'],
+        broker=app.config['CELERY_BROKER_URL'])
+    celery.conf.update(app.config)
+
+    TaskBase = celery.Task
+
+    class ContextTask(TaskBase):
+        abstract = True
+
+        def __call__(self, *args, **kwargs):
+            with app.app_context():
+                return TaskBase.__call__(self, *args, **kwargs)
+
+    celery.Task = ContextTask
+    return celery
